Allow installing without the app

+++ customize.sh
@@ -30,10 +30,14 @@ ui_print "- Extracting module files"
 
 unzip -o "$ZIPFILE" 'overlays/*' 'system/*' 'common/*' 'module.prop' 'system.prop' 'sepolicy.rule' 'zipsigner*' 'uninstall.sh' 'quickswitch' -d $MODPATH >&2
 chmod +x $MODPATH/common/*
-unzip -o "$ZIPFILE" 'QuickSwitch.apk' -d /data/local/tmp >&2
-ui_print "- installing QuickSwitch.apk"
-pm install -r "/data/local/tmp/QuickSwitch.apk"
-rm -rf /data/local/tmp/QuickSwitch.apk
+
+if [ ! -z "$NOAPK" ]; then
+  unzip -o "$ZIPFILE" 'QuickSwitch.apk' -d /data/local/tmp >&2
+  ui_print "- installing QuickSwitch.apk"
+  pm install -r "/data/local/tmp/QuickSwitch.apk"
+  rm -rf /data/local/tmp/QuickSwitch.apk
+fi
+
 [ "$($MODPATH/common/aaptx86 v)" ] && AAPT=aaptx86
 [ "$($MODPATH/common/aapt v)" ] && AAPT=aapt
 [ "$($MODPATH/common/aapt64 v)" ] && AAPT=aapt64
